third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/cross-origin-isolated-permission-worker.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] shared_worker (withCoopCoep: true) cross origin isolated permission test
  assert_equals: expected true but got false
Harness: the test ran to completion.
